chore: refactor api
