+++ src/api/workspace/config.rs
@@ -5,9 +5,9 @@ use std::{
 
 use crate::{
     api::WorkspaceApi,
-    cli::{ConfigFormat, ConfigPart, WorkEnvParams, WorkParams},
+    cli::{ConfigFormat, ConfigPart},
     config::{
-        config::{ConfigSource, FileFormat, RoozCfg},
+        config::{FileFormat, RoozCfg},
         runtime::RuntimeConfig,
     },
     model::{types::AnyError, volume::WORK_ROLE},
@@ -17,13 +17,8 @@ use crate::{
 use age::x25519::Identity;
 use colored::Colorize;
 
-pub enum UpdateMode {
-    Apply,
-    Purge,
-}
-
 impl<'a> WorkspaceApi<'a> {
-    pub async fn show_config(
+    pub async fn show(
         &self,
         workspace_key: &str,
         part: ConfigPart,
@@ -88,7 +83,7 @@ impl<'a> WorkspaceApi<'a> {
         io::stdin().read_line(&mut String::new()).unwrap();
     }
 
-    async fn edit_config_core(
+    pub async fn edit_config_core(
         &self,
         body: String,
         format: FileFormat,
@@ -130,70 +125,6 @@ impl<'a> WorkspaceApi<'a> {
         }
         edited_config.encrypt(identity).await?;
         Ok((edited_config, edited_body))
-    }
-
-    pub async fn update(
-        &self,
-        workspace_key: &str,
-        spec: &WorkEnvParams,
-        interactive: bool,
-        mode: UpdateMode,
-        no_pull: bool,
-    ) -> Result<(), AnyError> {
-        let labels = Labels::new(Some(workspace_key), Some(WORK_ROLE));
-
-        let container = self
-            .api
-            .container
-            .get_single(&labels)
-            .await?
-            .ok_or(format!("Workspace not found: {}", &workspace_key))?;
-
-        match mode {
-            UpdateMode::Apply => self.remove_containers_only(&workspace_key, true).await?,
-            UpdateMode::Purge => self.remove(&workspace_key, true).await?,
-        };
-
-        let identity = self.read_age_identity().await?;
-
-        if let Some(labels) = &container.labels {
-            if interactive {}
-
-            let config_source = &labels[labels::CONFIG_ORIGIN];
-            let format = FileFormat::from_path(config_source);
-            let original_body = &labels[labels::CONFIG_BODY];
-            let mut original_config = RoozCfg::deserialize_config(original_body, format)?.unwrap();
-
-            let config_to_apply = if interactive {
-                original_config.decrypt(&identity).await?;
-
-                let decrypted_string = original_config.to_string(format)?;
-                let (encrypted_config, _) = self
-                    .edit_config_core(decrypted_string.clone(), format, &identity)
-                    .await?;
-                encrypted_config
-            } else {
-                original_config
-            };
-
-            self.new(
-                &labels[labels::WORKSPACE_KEY],
-                &WorkParams {
-                    env: spec.clone(),
-                    pull_image: if no_pull { false } else { true },
-                    ..Default::default()
-                },
-                Some(ConfigSource::Body {
-                    value: config_to_apply,
-                    origin: config_source.to_string(),
-                    format,
-                }),
-                false,
-                &identity,
-            )
-            .await?;
-        }
-        Ok(())
     }
 
     pub async fn config_template(&self, _format: FileFormat) -> Result<(), AnyError> {
