+++ src/api/workspace/start.rs
@@ -2,7 +2,7 @@ use crate::{api::WorkspaceApi, model::types::AnyError, util::labels::Labels};
 use colored::Colorize;
 
 impl<'a> WorkspaceApi<'a> {
-    pub async fn start_workspace(&self, workspace_key: &str) -> Result<(), AnyError> {
+    pub async fn start(&self, workspace_key: &str) -> Result<(), AnyError> {
         let labels = Labels::new(Some(workspace_key), None);
 
         for c in self.api.container.get_all(&labels).await? {
