+++ src/cmd/mod.rs
@@ -3,3 +3,4 @@ pub mod list;
 pub mod new;
 pub mod prune;
 pub mod remote;
+pub mod update;
